requirements.txt
pip~=23.2.1
wheel~=0.41.2
pytz~=2023.3.post1
setuptools~=68.2.0
Django~=4.2.7
asgiref~=3.7.2
sqlparse~=0.4.4
requests~=2.31.0
certifi~=2023.11.17
idna~=3.6
urllib3~=2.1.0
django-filter~=23.4
django-extensions
django-cors-headers
djangorestframework~=3.14.0
python-dotenv~=1.0.0